pyproject.toml
[project]
name = "cinema"
version = "0.1.0"
description = "Cinema API backend"
authors = [
    { name = "youcef", email = "youcef.tabellout@outlook.fr" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
    "django>=5.2.7,<6.0.0",
    "djangorestframework>=3.16.1,<4.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "gunicorn>=23.0.0,<24.0.0",
    "django-environ (>=0.12.0,<0.13.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "httpx (>=0.28.1,<0.29.0)",
    "django-filter (>=25.2,<26.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^7.0.0"
pydocstyle = "^6.3.0"
ruff = "^0.14.3"

mypy = "^1.13.0"
django-stubs = "^5.1.1"
djangorestframework-stubs = "^3.15.0"

bandit = "^1.7.8"
types-django-filter = "^25.2.0.20251010"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
# Auto-fix safe issues
fix = true

line-length = 88

# Select includes:
# - E/F (pycodestyle/flake8)
# - B (bugbear)
# - C (complexity)
# - I (imports)
# - UP (pyupgrade)
select = ["E", "F", "B", "C", "I", "UP"]

# Ignore long lines (handled by black) and missing docstrings for now
ignore = ["E501", "D100", "D104", "D107"]

target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
plugins = ["mypy_django_plugin.main"]

# mypy plugin configuration for Django

[tool.django-stubs]
django_settings_module = "cinema.settings"
strict_settings = true

[tool.pydocstyle]
convention = "google"
add-ignore = ["D104", "D107", "D100"]
